server/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
PyMySQL==1.1.1
cryptography==42.0.8
python-dotenv==1.0.1